+++ apps/backend/src/app/audit-log/audit-log.field-labels.ts
@@ -1,0 +1,93 @@
+const FIELD_LABELS: Record<string, string> = {
+  academicId: 'Matrícula',
+  additionalPaymentInfo: 'Informações adicionais de pagamento',
+  allowSubscription: 'Permitir inscrição',
+  amountPaid: 'Valor pago',
+  attendedAt: 'Data da presença',
+  autoSubscribe: 'Inscrição automática',
+  buttonLink: 'Link do botão',
+  buttonText: 'Texto do botão',
+  category: 'Categoria',
+  contactInfo: 'Contato',
+  contactType: 'Tipo de contato',
+  committedById: 'Enviado por',
+  createdById: 'Autor',
+  createdByMethod: 'Origem',
+  deletedAt: 'Exclusão',
+  description: 'Descrição',
+  desiredCourses: 'Minicursos desejados',
+  desiredLectures: 'Palestras desejadas',
+  desiredUncategorized: 'Outros eventos desejados',
+  displayName: 'Nome público',
+  email: 'E-mail',
+  emoji: 'Emoji',
+  endDate: 'Fim',
+  eventGroupId: 'Grupo de eventos',
+  eventId: 'Evento',
+  externalRef: 'Referência externa',
+  identityDocument: 'Documento',
+  isActive: 'Ativa',
+  isPaymentRequired: 'Exigir pagamento',
+  latitude: 'Latitude',
+  locationDescription: 'Local',
+  longitude: 'Longitude',
+  majorEventId: 'Grande evento',
+  maxCoursesPerAttendee: 'Máximo de minicursos',
+  maxLecturesPerAttendee: 'Máximo de palestras',
+  maxUncategorizedPerAttendee: 'Máximo de outros eventos',
+  mergedIntoId: 'Unificada em',
+  name: 'Nome',
+  onlineAttendanceCode: 'Código de presença online',
+  onlineAttendanceEndDate: 'Fim da presença online',
+  onlineAttendanceStartDate: 'Início da presença online',
+  paymentDate: 'Data de pagamento',
+  paymentTier: 'Faixa de pagamento',
+  permission: 'Permissão',
+  personId: 'Pessoa',
+  phone: 'Telefone',
+  publiclyVisible: 'Visível publicamente',
+  publicationState: 'Estado da publicação',
+  publishedAt: 'Publicado em',
+  scheduledPublishAt: 'Publicação agendada',
+  unpublishedAt: 'Despublicado em',
+  publicationScheduledBy: 'Agendado por',
+  publicationUpdatedBy: 'Publicação atualizada por',
+  rankedSubscriptionEnabled: 'Inscrição por voto preferencial',
+  receiptRejectionReason: 'Motivo de rejeição',
+  receiptValidatedAt: 'Data da validação',
+  receiptValidatedBy: 'Validador',
+  scope: 'Escopo',
+  secondaryEmails: 'E-mails secundários',
+  selectedEventIds: 'Eventos selecionados',
+  shortDescription: 'Descrição curta',
+  shouldCollectAttendance: 'Coletar presença',
+  shouldIssueCertificate: 'Emitir certificado',
+  shouldIssueCertificateForEachEvent: 'Um certificado por evento',
+  shouldIssueCertificateForNonPayingAttendees: 'Certificado para não pagantes',
+  shouldIssueCertificateForNonSubscribedAttendees: 'Certificado para não inscritos',
+  shouldIssuePartialCertificate: 'Permitir certificado parcial',
+  shouldProvideSubscriberListToLecturer: 'Lista de inscritos para ministrante',
+  slots: 'Vagas',
+  startDate: 'Início',
+  subscriptionEndDate: 'Fim das inscrições',
+  subscriptionStartDate: 'Início das inscrições',
+  subscriptionStatus: 'Status da inscrição',
+  targetLabel: 'Alvo',
+  type: 'Tipo',
+  updatedById: 'Atualizado por',
+  userId: 'Usuário',
+  validFrom: 'Válida a partir de',
+  validUntil: 'Válida até',
+  whatsapp: 'WhatsApp',
+  youtubeCode: 'YouTube',
+};
+
+export function getAuditFieldLabel(field: string): string {
+  const exact = FIELD_LABELS[field];
+  if (exact) {
+    return exact;
+  }
+
+  const leaf = field.split('.').at(-1) ?? field;
+  return FIELD_LABELS[leaf] ?? leaf.replace(/([A-Z])/g, ' $1').replace(/^./, (letter) => letter.toUpperCase());
+}
